requirements.txt
gitasgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.0
tzdata==2024.1
cffi==1.16.0
cryptography==42.0.8
pycparser==2.22
django-rest-authtoken==2.1.4